su/lib cleanup
Former-commit-id: 6cd2da6e8071ff75fe508c47cb857cdc4d2898fb

+++ su/main/makevel.c
@@ -221,11 +221,11 @@ main (int argc, char **argv)
 
 				/* chirp */
 				if (z>z1c && z<z2c) {
-					vtemp = sin((ac+bc*z)*z);
+					vtemp = sinf((ac+bc*z)*z);
 					if (vtemp<0.0)
-						v[iz] -= vzc*pow(-vtemp,exc);
+						v[iz] -= vzc*powf(-vtemp,exc);
 					else
-						v[iz] += vzc*pow(vtemp,exc);
+						v[iz] += vzc*powf(vtemp,exc);
 				}
 
 				/* random perturbation */
